backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
dependencies = [
    "ml",
    "anyio",
    "annotated-doc",
    "starlette",
    "fastapi",
    "deprecated==1.3.1",
    "limits==5.8.0",
    "slowapi==0.1.9",
    "wrapt==2.1.2",
    "greenlet==3.3.0",
    "sqlalchemy==2.0.45",
    "psycopg==3.3.2",
    "psycopg-binary==3.3.2",
    "click==8.3.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "typer==0.21.1",
    "shellingham==1.5.4",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "httptools==0.7.1",
    "websockets==16.0",
    "dnspython==2.8.0",
    "email-validator==2.3.0",
    "fastapi-cli==0.0.20",
    "fastapi-cloud-cli==0.9.0",
    "fastar==0.8.0",
    "python-multipart==0.0.21",
    "pydantic-extra-types==2.11.0",
    "celery==5.6.2",
    "amqp==5.3.1",
    "billiard==4.2.4",
    "kombu==5.6.2",
    "vine==5.1.0",
    "tzlocal==5.3.1",
    "redis==7.1.0",
    "rich==14.2.0",
    "rich-toolkit==0.17.1",
    "prompt-toolkit==3.0.52",
    "wcwidth==0.2.14",
    "pygments==2.19.2",
    "markdown-it-py==4.0.0",
    "mdurl==0.1.2",
    "uvicorn==0.40.0",
    "uvloop==0.22.1",
    "watchfiles==1.1.1",
    "sentry-sdk==2.49.0",
    "rignore==0.7.6"
]

[tool.setuptools.packages.find]
where = ["."]